requirements.txt
chromadb
langchain
streamlit

cryptography
pandas
pypdf
python-dotenv

langchain_chroma
langchain_community
langchain_ollama
langchain_text_splitters
